refactor: 유틸리티 타입(Omit)을 사용해 중복되는 타입 선언 제거
