+++ src/types/mission/MissionItem.ts
@@ -1,0 +1,8 @@
+export interface MissionItem {
+    missionId: number;
+    missionName: string;
+    missionOrder: number;
+    completed: boolean;
+    isEditing: boolean;
+    isChecked: boolean;
+}
